P4-3091 moving Event, Person and Profile domain entities into the Domain level package.
