+++ src/main/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/domain/move/Profile.kt
@@ -1,7 +1,9 @@
-package uk.gov.justice.digital.hmpps.pecs.jpc.service.spreadsheet.inbound.report
+package uk.gov.justice.digital.hmpps.pecs.jpc.domain.move
 
 import com.beust.klaxon.Json
 import com.beust.klaxon.Klaxon
+import uk.gov.justice.digital.hmpps.pecs.jpc.service.spreadsheet.inbound.report.EventDateTime
+import uk.gov.justice.digital.hmpps.pecs.jpc.service.spreadsheet.inbound.report.dateTimeConverter
 import java.time.LocalDateTime
 import javax.persistence.Column
 import javax.persistence.Entity
